+++ src/renderer/renderer2d/shaders/sprite_vs.glsl
@@ -1,7 +1,8 @@
 layout(location = 0) in vec3 vertexPosition;
 layout(location = 1) in vec2 vertexTexCoord;
 layout(location = 2) in vec4 vertexColor;
-layout(location = 3) in mat4 worldMatrix;
+layout(location = 3) in vec4 uvTransform;
+layout(location = 4) in mat4 worldMatrix;
 
 uniform mat4 viewProjection;
 
@@ -11,7 +12,7 @@ out vec3 fragmentPosition;
 
 void main()
 {
-    texCoord = vertexTexCoord;
+    texCoord = vertexTexCoord * uvTransform.zw + uvTransform.xy;
     vec4 worldPosition = worldMatrix * vec4(vertexPosition, 1.0);
     fragmentPosition = worldPosition.xyz;
     gl_Position = viewProjection * worldPosition;